CXBuddyPro/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
openai==1.12.0
aiohttp==3.9.1
chromadb==0.4.22
numpy<2.0.0
azure-cognitiveservices-speech==1.34.0
